Android 13.0.0 release 35

* tag 'android-13.0.0_r35': (667 commits)
  Revert "Fix an OOB bug in bta_hh_co_get_rpt_rsp"
  Revert "Don't send MAP new message events if older than one year"
  Harden array bounds validation
  Fix an OOB write in SDP_AddAttribute
  Fix an OOB bug in bta_hh_co_get_rpt_rsp
  Fix an OOB access bug in A2DP_BuildMediaPayloadHeaderSbc
  Fix an OOB Write bug in gatt_check_write_long_terminate
  Revert "Set connection state to not connected while ACL is lost"
  Don't send MAP new message events if older than one year
  Add a few BluetoothMapServiceTest methods
  Add PanServiceTest methods
  Add tests for BluetoothMapMasInstance and BluetoothMapObexServer
  Add additional GattServiceTest
  Add tests for handle methods in BluetoothMapContentObserver
  Add tests for message getters in BluetoothMapContent
  change BAS connectGatt from TRANSPORT_AUTO to TRANSPORT_LE
  leaudio: Fix reconnection
  LE Audio: improve LE Audio device detection for Bonding
  csip: Trigger active scan when read group size
  leaudio: Add more logs for TA in dumpsys
  ...

+++ system/embdrv/encoder_for_aptx/src/aptXbtenc.c
@@ -1,3 +1,18 @@
+/**
+ * Copyright (C) 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 #include "aptXbtenc.h"
 
 #include "AptxEncoder.h"
@@ -200,13 +215,13 @@ APTXBTENCEXPORT int aptxbtenc_encodestereo(void* _state, void* _pcmL,
   tmp_out = tmp_reg >> state->m_endian;
   tmp_out |= tmp_reg << state->m_endian;
 
-  buffer[0] = (int16_t)tmp_out;
+  buffer[0] = tmp_out;
   tmp_reg = packCodeword(&state->m_encoderData[1], state->m_sync_mode);
   // Swap bytes to output data in big-endian as expected by bc5 code...
   tmp_out = tmp_reg >> state->m_endian;
   tmp_out |= tmp_reg << state->m_endian;
 
-  buffer[1] = (int16_t)tmp_out;
+  buffer[1] = tmp_out;
 
   return 0;
 }
